+++ packages/menu/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @launchpad-ui/menu
 
+## 0.13.4
+
+### Patch Changes
+
+- Updated dependencies [[`31b5c98`](https://github.com/launchdarkly/launchpad-ui/commit/31b5c98c49768e62f6474c8696b40e9cb52ea25d)]:
+  - @launchpad-ui/tokens@0.9.6
+  - @launchpad-ui/form@0.11.4
+  - @launchpad-ui/icons@0.16.3
+  - @launchpad-ui/popover@0.11.15
+  - @launchpad-ui/tooltip@0.9.4
+
 ## 0.13.3
 
 ### Patch Changes