+++ packages/snackbar/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @launchpad-ui/snackbar
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies [[`31b5c98`](https://github.com/launchdarkly/launchpad-ui/commit/31b5c98c49768e62f6474c8696b40e9cb52ea25d)]:
+  - @launchpad-ui/tokens@0.9.6
+  - @launchpad-ui/button@0.12.3
+  - @launchpad-ui/icons@0.16.3
+
 ## 0.5.2
 
 ### Patch Changes